pyproject.toml
[project]
name = "automagik-hive"
version = "0.2.0rc16"
description = "Automagik Hive"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
 "agno==1.7.5",
 "aiofiles>=24.1.0",
 "jinja2>=3.1.6",
 "markdown>=3.8.2",
 "watchdog>=6.0.0",
 "psutil>=5.9.0",
 "asyncpg>=0.29.0",
 "pgvector>=0.2.0",
 "alembic>=1.16.4",
 "psycopg[binary]>=3.1.0",
 "psycopg-pool>=3.1.0",
 "sqlalchemy>=2.0.0",
 "anthropic>=0.31.0",
 "openai>=1.93.2",
 "cohere>=5.15.0",
 "google-generativeai>=0.8.5",
 "google-genai>=1.25.0",
 "fastapi>=0.116.0",
 "uvicorn>=0.35.0",
 "aiohttp>=3.12.14",
 "requests>=2.32.4",
 "mcp>=1.10.1",
 "loguru>=0.7.3",
 "langwatch>=0.2.9",
 "openinference-instrumentation-agno>=0.1.9",
 "pandas>=2.3.1",
 "docker>=7.1.0",
 "psycopg2-binary>=2.9.10",
 "google-api-python-client>=2.176.0",
 "google-auth-httplib2>=0.2.0",
 "google-auth-oauthlib>=1.2.2",
 "openpyxl>=3.1.0",
 "pyxlsb>=1.0.10",
 "xlrd>=2.0.2",
 "odfpy>=1.4.1",
 "python-dateutil>=2.8.0",
 "pypdf>=6.1.1",
]
[[project.authors]]
name = "Automagik"
email = "hive@namastex.ai"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["ai", "api", "lib", "cli", "common"]

[tool.hatch.build.targets.wheel.shared-data]
"ai/agents/template-agent" = "automagik_hive/templates/agents/template-agent"
"ai/teams/template-team" = "automagik_hive/templates/teams/template-team"
"ai/workflows/template-workflow" = "automagik_hive/templates/workflows/template-workflow"
".env.example" = "automagik_hive/templates/.env.example"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pytest-asyncio>=1.1.0",
    "coverage>=7.10.4",
    "pytest-mock>=3.14.1",
    "psutil>=7.0.0",
    "twine>=6.1.0",
    "pytest-cov>=6.2.1",
    "pandas-stubs>=2.3.2.250827",
    "types-pyyaml>=6.0.12.20250915",
    "types-tqdm>=4.67.0.20250809",
]

[tool.ruff]
line-length = 120
target-version = "py312"
exclude = [
      ".claude",
      ".git",
      ".venv",
      "__pycache__",
      "build",
      "dist",
  ]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "S", "B", "A", "C4", "T20"]
ignore = ["E501", "S101", "S603", "S607", "T201", "B904", "S106", "B008"]


[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
exclude = [
      '^\.venv/',
      '^\.git/',
      '__pycache__',
]
[[tool.mypy.overrides]]
  module = ["langwatch.*", "langwatch.dspy.*"]
  follow_imports = "skip"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
